% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basketballpullr-package.R
\docType{package}
\name{basketballpullr-package}
\alias{basketballpullr}
\alias{basketballpullr-package}
\title{basketballpullr: Pull Basketball Data from Various Sources}
\description{
System for simplifying the gathering of basketball statistics and data. Focuses on variety of statistic types, leagues, and sources.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jcampbellsjci/basketballpullr}
  \item Report bugs at \url{https://github.com/jcampbellsjci/basketballpullr/issues}
}

}
\author{
\strong{Maintainer}: Jake Campbell \email{jcampbellsjci@gmail.com}

}
\keyword{internal}
