% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_season_table.R
\name{pull_season_table}
\alias{pull_season_table}
\title{Pull season tables}
\usage{
pull_season_table(season, table)
}
\arguments{
\item{season}{Season the stats were collected in.}

\item{table}{Specific statistic table to be pulled.}
}
\value{
A dataframe.
}
\description{
Pull season tables
}
\examples{
pull_season_table(season = 2021, table = "per_game")
}
